+++ src/test/java/TestAliceInWonderland.kt
@@ -29,10 +29,12 @@ import com.planbase.pdf.layoutmanager.pages.SinglePage
 import com.planbase.pdf.layoutmanager.utils.CMYK_BLACK
 import com.planbase.pdf.layoutmanager.utils.Coord
 import com.planbase.pdf.layoutmanager.utils.Dim
+import org.apache.pdfbox.cos.COSString
 import org.apache.pdfbox.pdmodel.common.PDRectangle
 import org.apache.pdfbox.pdmodel.font.PDType0Font
 import org.apache.pdfbox.pdmodel.font.PDType1Font
 import org.apache.pdfbox.pdmodel.graphics.color.PDDeviceCMYK
+import org.apache.pdfbox.util.Charsets
 import org.junit.Test
 import java.io.File
 import java.io.FileOutputStream
@@ -51,11 +53,15 @@ class TestAliceInWonderland {
         val titleFontFile = File("target/test-classes/EmilysCandy-Regular.ttf")
         val titleFont: PDType0Font = pageMgr.loadTrueTypeFont(titleFontFile)
 
+        val handwritingFontFile = File("target/test-classes/MrDeHaviland-Regular.ttf")
+        val handwritingFont: PDType0Font = pageMgr.loadTrueTypeFont(handwritingFontFile)
+
         val incipit = TextStyle(titleFont, 36.0, CMYK_BLACK, 30.0)
         val chapTitleCellStyle = CellStyle(BOTTOM_LEFT, BoxStyle(Padding(50.0, 0.0, 10.0, 0.0), null, NO_BORDERS))
         val bodyCellStyle = CellStyle(TOP_LEFT_JUSTIFY, BoxStyle(Padding(10.0, 0.0, 0.0, 0.0), null, NO_BORDERS))
         val heading = TextStyle(titleFont, 16.0, CMYK_BLACK, 16.0, 0.0, 0.2, 0.0)
         val bodyText = TextStyle(PDType1Font.TIMES_ROMAN, 12.0, CMYK_BLACK)
+        val handwriting = TextStyle(handwritingFont, 17.0, CMYK_BLACK, 20.0, wordSpacing = 4.0)
         val thought = TextStyle(PDType1Font.TIMES_ITALIC, 12.0, CMYK_BLACK)
 
         val lp = pageMgr.startPageGrouping(
@@ -76,13 +82,13 @@ class TestAliceInWonderland {
         val index:TreeMap<String,MutableList<IntRange>> = TreeMap()
 
         var dap: DimAndPageNums =
-                lp.appendCell(chapTitleCellStyle,
-                            listOf(Text(incipit, "1. "),
-                                   Text(heading, "Down the Rabbit Hole")))
+                lp.appendCell(0.0, chapTitleCellStyle,
+                              listOf(Text(incipit, "1. "),
+                                     Text(heading, "Down the Rabbit Hole")))
         assertEquals(IntRange(1, 1), dap.pageNums)
         tableOfContents.add(TitlePage("1. Down the Rabbit Hole", dap.pageNums.start))
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "Alice was beginning to get very tired of sitting by her sister on" +
                                         " the bank, and of having nothing to do: once or twice she had" +
@@ -94,7 +100,7 @@ class TestAliceInWonderland {
         assertEquals(IntRange(1, 1), dap.pageNums)
         addToIndex("Bank", dap, index)
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "So she was considering, in her own mind (as well as she could, for" +
                                         " the hot day made her feel very sleepy and stupid), whether the" +
@@ -104,7 +110,7 @@ class TestAliceInWonderland {
         assertEquals(IntRange(1, 1), dap.pageNums)
         addToIndex("White Rabbit", dap, index)
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "There was nothing so very remarkable in that; nor did Alice think" +
                                         " it so very much out of the way to hear the Rabbit say to itself," +
@@ -121,7 +127,7 @@ class TestAliceInWonderland {
         addToIndex("Feet", dap, index)
         addToIndex("Waistcoat-pocket", dap, index)
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "In another moment down went Alice after it, never once considering how" +
                                         " in the world she was to get out again....")))
@@ -130,13 +136,13 @@ class TestAliceInWonderland {
         // Forces a new page.
         lp.cursorToNewPage()
 
-        dap = lp.appendCell(chapTitleCellStyle,
-                              listOf(Text(incipit, "2. "),
-                                     Text(heading, "The Pool of Tears")))
+        dap = lp.appendCell(0.0, chapTitleCellStyle,
+                            listOf(Text(incipit, "2. "),
+                                   Text(heading, "The Pool of Tears")))
         assertEquals(IntRange(3, 3), dap.pageNums)
         tableOfContents.add(TitlePage("2. The Pool of Tears", dap.pageNums.start))
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "\"Curiouser and curiouser!\" cried Alice (she was so much surprised, that" +
                                         " for the moment she quite forgot how to speak good English). \"Now I'm" +
@@ -151,38 +157,44 @@ class TestAliceInWonderland {
         assertEquals(IntRange(3, 3), dap.pageNums)
         addToIndex("Feet", dap, index)
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "And she went on planning to herself how she would manage it. "),
                                    Text(thought, "They must go by the carrier, "),
                                    Text(bodyText, "she thought; "),
                                    Text(thought,
                                         "and how funny it’ll seem, sending presents to one’s own feet!" +
-                                        " And how odd the directions will look!\n" +
-                                        "\n" +
+                                        " And how odd the directions will look!")))
+        assertEquals(IntRange(4, 4), dap.pageNums)
+        addToIndex("Feet", dap, index)
+
+        dap = lp.appendCell(40.0, bodyCellStyle,
+                            listOf(Text(handwriting,
                                         "Alice’s Right Foot, Esq.\n" +
-                                        "Hearthrug,\n" +
-                                        "near the Fender,\n" +
-                                        "(with Alice’s love).\n" +
-                                        "\n" +
+                                        "Hearthrug\n" +
+                                        "Near the Fender\n" +
+                                        "(with Alice’s love)")))
+        assertEquals(IntRange(4, 4), dap.pageNums)
+
+        dap = lp.appendCell(0.0, bodyCellStyle,
+                            listOf(Text(thought,
                                         "Oh dear, what nonsense I’m talking!\n\n"),
                                    Text(bodyText,
                                         "Just then her head struck against the roof of the hall: in fact she was" +
                                         " now more than nine feet high, and she at once took up the little golden" +
                                         " key and hurried off to the garden door....")))
         assertEquals(IntRange(4, 4), dap.pageNums)
-        addToIndex("Feet", dap, index)
 
         lp.cursorToNewPage()
 
         // TODO: Here there's no space between "a" and "Long".  I thought that bug was fixed!
-        dap = lp.appendCell(chapTitleCellStyle,
-                             listOf(Text(incipit, "3. "),
-                                    Text(heading, "A Caucus-Race and a Long Tale")))
+        dap = lp.appendCell(0.0, chapTitleCellStyle,
+                            listOf(Text(incipit, "3. "),
+                                   Text(heading, "A Caucus-Race and a Long Tale")))
         assertEquals(IntRange(5, 5), dap.pageNums)
         tableOfContents.add(TitlePage("3. A Caucus-Race and a Long Tale", dap.pageNums.start))
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "They were indeed a queer-looking party that assembled on the bank—the" +
                                         " birds with draggled feathers, the animals with their fur clinging close" +
@@ -190,7 +202,7 @@ class TestAliceInWonderland {
         assertEquals(IntRange(5, 5), dap.pageNums)
         addToIndex("Bank", dap, index)
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "The first question of course was, how to get dry again: they had a" +
                                         " consultation about this, and after a few minutes it seemed quite" +
@@ -203,7 +215,7 @@ class TestAliceInWonderland {
         assertEquals(IntRange(5, 6), dap.pageNums)
         addToIndex("Lory, the", dap, index)
 
-        dap = lp.appendCell(bodyCellStyle,
+        dap = lp.appendCell(0.0, bodyCellStyle,
                             listOf(Text(bodyText,
                                         "At last the Mouse, who seemed to be a person of authority among them," +
                                         " called out, \"Sit down, all of you, and listen to me! I’ll soon make" +
@@ -214,8 +226,8 @@ class TestAliceInWonderland {
 
         lp.cursorToNewPage()
 
-        dap = lp.appendCell(chapTitleCellStyle,
-                             listOf(Text(heading, "Index")))
+        dap = lp.appendCell(0.0, chapTitleCellStyle,
+                            listOf(Text(heading, "Index")))
         tableOfContents.add(TitlePage("Index", dap.pageNums.start))
 
         for (item in index) {
@@ -234,7 +246,7 @@ class TestAliceInWonderland {
                 }))
             }
 
-            lp.appendCell(bodyCellStyle, ret)
+            lp.appendCell(0.0, bodyCellStyle, ret)
         }
 
         var sp = SinglePage(-2, pageMgr, null, lp.body)
@@ -265,12 +277,16 @@ class TestAliceInWonderland {
         assertEquals(IntRange(-1, -1), dap.pageNums)
         pageMgr.insertPageAt(sp, 1)
 
-        pageMgr.commit()
-        // We're just going to write to a file.
-        val os = FileOutputStream("alice.pdf")
+        // If we don't set the file identifiers, PDFBox will set them for us.  We only care because we're checking
+        // this into git so that when something breaks, we can easily compare with a known-good file to see how big
+        // a change it is.
+        val docId = COSString("Alice In Wonderland Test PDF".toByteArray(Charsets.ISO_8859_1))
+        pageMgr.setFileIdentifiers(docId, docId)
 
-        // Commit it to the output stream!
-        pageMgr.save(os)
+        pageMgr.commit()
+
+        // We're just going to write to a file.
+        pageMgr.save(FileOutputStream("alice.pdf"))
     }
 
     private fun addToIndex(