+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 2018-05-08
+ - Changed all Floats to Doubles
+ - Aded PdfLayoutMgr.getFileIdentifiers() and .setFileIdentifiers()
+ - Added handwriting font
+ - Started saving PDF files so that we can see differences.
+ - Changed PageGrouping.appendCell() to take an x-offset as the first parameter (usually set to zero)
+
 # 2.0.10 - 2018-03-29 "Fill-Rect after many pages"
  - Fixed bug where after a hundred or more pages, the edge of a fill-rect in a PageGrouping could slip between the
  floating point gaps and cause an exception.